+++ blakserv/ccode.c
@@ -113,7 +113,7 @@ const char* stristr(const char* pSource, const char* pSearch)
    if (!pSource || !pSearch || !*pSearch)
       return NULL;
 	
-   int nSearch = strlen(pSearch);
+   size_t nSearch = strlen(pSearch);
    // Don't search past the end of pSource
    const char *pEnd = pSource + strlen(pSource) - nSearch;
    while (pSource <= pEnd)
@@ -389,7 +389,7 @@ int C_GodLog(int object_id,local_var_type *local_vars,
 			
 		case TAG_STRING :
 			{
-				int lenBuffer;
+				size_t lenBuffer;
 				string_node *snod = GetStringByID(each_val.v.data);
 				
 				if (snod == NULL)
@@ -405,7 +405,7 @@ int C_GodLog(int object_id,local_var_type *local_vars,
 			
 		case TAG_TEMP_STRING :
 			{
-				int len_buf;
+				size_t len_buf;
 				string_node *snod;
 				
 				snod = GetTempString();
@@ -532,7 +532,7 @@ int C_Debug(int object_id,local_var_type *local_vars,
 			
 		case TAG_STRING :
 			{
-				int lenBuffer;
+				size_t lenBuffer;
 				string_node *snod = GetStringByID(each_val.v.data);
 				
 				if (snod == NULL)
@@ -548,7 +548,7 @@ int C_Debug(int object_id,local_var_type *local_vars,
 			
 		case TAG_TEMP_STRING :
 			{
-				int len_buf;
+				size_t len_buf;
 				string_node *snod;
 				
 				snod = GetTempString();
@@ -1200,7 +1200,7 @@ bool LookupString(val_type val, const char *function_name, const char **str, int
 
    if (*str == NULL)
       return false;
-   *len = strlen(*str);
+   *len = (int)strlen(*str);
    
    return true;
 }
@@ -1274,8 +1274,8 @@ int C_StringEqual(int object_id,local_var_type *local_vars,
          return ret_val.int_val;
       }
 
-      len1 = strlen(s1);
-      len2 = strlen(s2);
+      len1 = (int)strlen(s1);
+      len2 = (int)strlen(s2);
       ret_val.v.data = FuzzyBufferEqual(s1, len1, s2, len2);
 
       return ret_val.int_val;
@@ -1284,7 +1284,7 @@ int C_StringEqual(int object_id,local_var_type *local_vars,
    // First string is resource, second isn't.
    if (s1_resource)
    {
-      len2 = strlen(s2);
+      len2 = (int)strlen(s2);
       for (int i = 0; i < MAX_LANGUAGE_ID; i++)
       {
          s1 = r1->resource_val[i];
@@ -1298,7 +1298,7 @@ int C_StringEqual(int object_id,local_var_type *local_vars,
             continue;
          }
 
-         len1 = strlen(s1);
+         len1 = (int)strlen(s1);
          if (FuzzyBufferEqual(s1, len1, s2, len2))
          {
             ret_val.v.data = True;
@@ -1311,7 +1311,7 @@ int C_StringEqual(int object_id,local_var_type *local_vars,
    // Second string is resource, first isn't.
    if (s2_resource)
    {
-      len1 = strlen(s1);
+      len1 = (int)strlen(s1);
       for (int i = 0; i < MAX_LANGUAGE_ID; i++)
       {
          s2 = r2->resource_val[i];
@@ -1325,7 +1325,7 @@ int C_StringEqual(int object_id,local_var_type *local_vars,
             continue;
          }
 
-         len2 = strlen(s2);
+         len2 = (int)strlen(s2);
          if (FuzzyBufferEqual(s1, len1, s2, len2))
          {
             ret_val.v.data = True;
@@ -1336,8 +1336,8 @@ int C_StringEqual(int object_id,local_var_type *local_vars,
    }
 
    // Neither strings are resources.
-   len1 = strlen(s1);
-   len2 = strlen(s2);
+   len1 = (int)strlen(s1);
+   len2 = (int)strlen(s2);
    ret_val.v.data = FuzzyBufferEqual(s1, len1, s2, len2);
    return ret_val.int_val;
 }
@@ -1477,7 +1477,7 @@ int C_StringSubstitute(int object_id,local_var_type *local_vars,
          bprintf( "C_StringSub can't sub for invalid resource %i\n", s1_val.v.data );
          return NIL;
       }
-      len1 = strlen(s1);
+      len1 = (int)strlen(s1);
       break;
 
    case TAG_DEBUGSTR :
@@ -1496,7 +1496,7 @@ int C_StringSubstitute(int object_id,local_var_type *local_vars,
       s1 = GetClassDebugStr(c,s1_val.v.data);
       len1 = 0;
       if (s1)
-         len1 = strlen(s1);
+         len1 = (int)strlen(s1);
       break;
 
    case TAG_NIL :
@@ -1634,8 +1634,8 @@ int C_StringContain(int object_id,local_var_type *local_vars,
          return ret_val.int_val;
       }
 
-      len1 = strlen(s1);
-      len2 = strlen(s2);
+      len1 = (int)strlen(s1);
+      len2 = (int)strlen(s2);
       ret_val.v.data = FuzzyBufferContain(s1, len1, s2, len2);
 
       return ret_val.int_val;
@@ -1644,7 +1644,7 @@ int C_StringContain(int object_id,local_var_type *local_vars,
    // First string is resource, second isn't.
    if (s1_resource)
    {
-      len2 = strlen(s2);
+      len2 = (int)strlen(s2);
       ret_val.v.tag = TAG_INT;
       for (int i = 0; i < MAX_LANGUAGE_ID; i++)
       {
@@ -1659,7 +1659,7 @@ int C_StringContain(int object_id,local_var_type *local_vars,
             continue;
          }
 
-         len1 = strlen(s1);
+         len1 = (int)strlen(s1);
          if (FuzzyBufferContain(s1, len1, s2, len2))
          {
             ret_val.v.data = True;
@@ -1672,7 +1672,7 @@ int C_StringContain(int object_id,local_var_type *local_vars,
    // Second string is resource, first isn't.
    if (s2_resource)
    {
-      len1 = strlen(s1);
+      len1 = (int)strlen(s1);
       ret_val.v.tag = TAG_INT;
       for (int i = 0; i < MAX_LANGUAGE_ID; i++)
       {
@@ -1687,7 +1687,7 @@ int C_StringContain(int object_id,local_var_type *local_vars,
             continue;
          }
 
-         len2 = strlen(s2);
+         len2 = (int)strlen(s2);
          if (FuzzyBufferContain(s1, len1, s2, len2))
          {
             ret_val.v.data = True;
@@ -1698,8 +1698,8 @@ int C_StringContain(int object_id,local_var_type *local_vars,
    }
 
    // Neither strings are resources.
-   len1 = strlen(s1);
-   len2 = strlen(s2);
+   len1 = (int)strlen(s1);
+   len2 = (int)strlen(s2);
    ret_val.v.data = FuzzyBufferContain(s1, len1, s2, len2);
    return ret_val.int_val;
 }
@@ -1761,7 +1761,7 @@ int C_SetResource(int object_id,local_var_type *local_vars,
 					str_val.v.data);
 				return NIL;
 			}
-			new_len = strlen(r->resource_val[0]);
+         new_len = (int)strlen(r->resource_val[0]);
 			new_str = r->resource_val[0];
 			break;
 		}
@@ -1869,7 +1869,7 @@ int C_ParseString(int object_id,local_var_type *local_vars,
 		us because we null terminated the real string*/
 		
 		strcpy(snod->data,each_str);
-		snod->len_data = strlen(snod->data);
+      snod->len_data = (int)strlen(snod->data);
 		
 		SendBlakodMessage(object_id,callback_val.v.data,1,p);
 		
@@ -1945,7 +1945,7 @@ int C_SetString(int object_id,local_var_type *local_vars,
          return NIL;
       }
       //bprintf("SetString string%i<--resource%i\n",s1_val.v.data,s2_val.v.data);
-      SetString(snod, (char*)str, strlen(str));
+      SetString(snod, (char*)str, (int)strlen(str));
       break;
 
    case TAG_MESSAGE :
@@ -1955,7 +1955,7 @@ int C_SetString(int object_id,local_var_type *local_vars,
          bprintf("C_SetString can't set from invalid message %i\n",s2_val.v.data);
          return NIL;
       }
-      SetString(snod,GetNameByID(s2_val.v.data),strlen(GetNameByID(s2_val.v.data)));
+      SetString(snod,GetNameByID(s2_val.v.data),(int)strlen(GetNameByID(s2_val.v.data)));
       break;
 
    case TAG_CLASS :
@@ -1972,7 +1972,7 @@ int C_SetString(int object_id,local_var_type *local_vars,
          return NIL;
       }
 
-      SetString(snod, c->class_name, strlen(c->class_name));
+      SetString(snod, c->class_name, (int)strlen(c->class_name));
       break;
 
    case TAG_DEBUGSTR :
@@ -1982,13 +1982,13 @@ int C_SetString(int object_id,local_var_type *local_vars,
          bprintf("C_SetString can't set from invalid debug string %i\n",s2_val.v.data);
          return NIL;
       }
-      SetString(snod,(char*)str,strlen(str));
+      SetString(snod,(char*)str,(int)strlen(str));
       break;
 
    case TAG_INT:
       char buf[21];
       snprintf(buf, 21, "%i", s2_val.v.data);
-      SetString(snod, buf, strlen(buf));
+      SetString(snod, buf, (int)strlen(buf));
       break;
 
    default :
@@ -2063,7 +2063,7 @@ int C_AppendTempString(int object_id,local_var_type *local_vars,
          bprintf("C_AppendTempString can't set from invalid resource %i\n", s_val.v.data);
          return NIL;
       }
-      AppendTempString(pStrConst, strlen(pStrConst));
+      AppendTempString(pStrConst, (int)strlen(pStrConst));
    }
       break;
 	case TAG_DEBUGSTR :
@@ -2082,7 +2082,7 @@ int C_AppendTempString(int object_id,local_var_type *local_vars,
 			strLen = 0;
 			if (pStrConst != NULL)
 			{
-				strLen = strlen(pStrConst);
+            strLen = (int)strlen(pStrConst);
 				AppendTempString(pStrConst,strLen);
 			}
 			else
